ovn: ovn-ctl support to start ovn db servers in backup mode
This patch adds support to start_ovsdb() function in ovn-ctl to start the
ovn db servers in backup mode. This can be done in the following ways
1. Use parameters --ovn-nb-sync-from-addr and --ovn-sb-sync-from-addr to
   set the addresses of the active server.
2. Create files $etcdir/ovnnb-active.conf and $etcdir/ovnsb-active.conf
   with the tcp url of the active servers.

If --ovn-nb-sync-from-addr and --ovn-sb-sync-from-addr is used, it will
overwrite the contents in the $etcdir/*.conf and use that server as the
active server.

Additional functions to promote a backup server to active and demote
active server to backup mode are also added in this patch

start_northd is changed such that it tries to connect to the db server urls
in $etcdir/*.conf when the *.conf files exist, else it will connect to
the unix socket.

Signed-off-by: Babu Shanmugam <bschanmu@redhat.com>