ovn: Cache mac bindings from broadcasted dynamic ARP responses
This patch attempts to avoid the usage of MAC_Binding table.

Dynamic ARP response originates from the logical ports with "unknown"
address. When the ARP resolution is requested via a logical
router datapath, ARP response will be delivered to the logical router
port of the switch. Since the logical router is distributed, the
response will never reach the other chassis from where the ARP resolution
is requested.

This is done using a OVN logical action bcast2lr() which adds an action
to send the packet to a specific multicast group for the logical router
datapath. For the chassis that have an "unknown" port connected to any
logical router, the bcast2lr() action will do the broadcast. The other
chassis will listen for such broadcast packets and executes put_arp()
action on them.

When the ovn-controller processes put_arp() action, it caches the MAC
bindings and this cache will be used to update the openflow table used
for the dynamic MAC resolution

Signed-off-by: Babu Shanmugam <bschanmu@redhat.com>